DHIS2-702: add support for Year On Year chart type (#4)
Year on year requires a different way of generating the series data, as
each response is 1 serie, while in the other chart types there could be
multiple serie in the same response (depending on the data selection).
The xAxis also needs to be customised to show the generic period label for the categories periods (ie. months per year). The labels currently come from the backend.